requirements.txt
Django==2.2.19
django-multiselectfield
Pillow
sorl-thumbnail
djangorestframework
django-debug-toolbar 